requirements.txt
pygame-ce>=2.3.1
numpy>=1.24.3